+++ src/Node.java
@@ -2,6 +2,6 @@ import java.util.Map;
 
 public abstract class Node {
 
-    public abstract int evaluate(Map<String, Integer> variableMap);
+    public abstract double evaluate(Map<String, Double> variableMap);
 
 }
